website/backend/requirements.txt
fastapi>=0.110,<1
uvicorn>=0.27,<1
Pillow>=10,<12
